ashpy/pyproject.toml
[project]
name = "ashpy"
version = "0.0.1"
description = "ash-code Python sidecar: skills, commands, LLM providers"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.66",
    "grpcio-tools>=1.66",
    "protobuf>=5.28",
    "pydantic>=2.8",
    "python-frontmatter>=1.1",
    "watchdog>=5.0",
    "jinja2>=3.1",
    "httpx>=0.27",
    "anthropic>=0.39",
    "openai>=1.54",
    "ollama>=0.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
]

[project.scripts]
ashpy = "ashpy.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ashpy"]

[tool.uv]
package = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
